[#67] Move startup processes into a loader
